+++ lib/src/model/asn1/pkcs/pkcs12_parameter_generator.dart
@@ -1,13 +1,57 @@
 import 'dart:typed_data';
 
-import 'package:pointycastle/api.dart';
+import 'package:pointycastle/pointycastle.dart';
 
-class Pkcs12HelperUtils {
-  static Uint8List generateDerivedKey(int idByte, int n, Uint8List salt,
-      Uint8List password, Digest digest, int iterationCount) {
-    var u = digest.digestSize;
-    var v = digest.byteLength;
+class PKCS12ParametersGenerator {
+  static final int KEY_MATERIAL = 1;
 
+  static final int IV_MATERIAL = 2;
+
+  static final int MAC_MATERIAL = 3;
+
+  Digest digest;
+
+  late int u;
+
+  late int v;
+
+  late Uint8List salt;
+
+  late Uint8List password;
+  late int iterationCount;
+
+  PKCS12ParametersGenerator(this.digest) {
+    u = digest.digestSize;
+    v = digest.byteLength;
+  }
+
+  void init(Uint8List password, Uint8List salt, int iterationCount) {
+    this.password = password;
+    this.salt = salt;
+    this.iterationCount = iterationCount;
+  }
+
+  KeyParameter generateDerivedParameters(int keySize) {
+    var dKey = _generateDerivedKey(KEY_MATERIAL, keySize);
+
+    return KeyParameter(dKey);
+  }
+
+  ParametersWithIV generateDerivedParametersWithIV(int keySize, int ivSize) {
+    var dKey = _generateDerivedKey(KEY_MATERIAL, keySize);
+
+    var iv = _generateDerivedKey(IV_MATERIAL, ivSize);
+
+    return ParametersWithIV(KeyParameter(dKey), iv);
+  }
+
+  KeyParameter generateDerivedMacParameters(int keySize) {
+    var dKey = _generateDerivedKey(MAC_MATERIAL, keySize);
+
+    return KeyParameter(dKey);
+  }
+
+  Uint8List _generateDerivedKey(int idByte, int n) {
     var D = Uint8List(v);
     var dKey = Uint8List(n);
     for (var i = 0; i != D.length; i++) {
@@ -32,8 +76,8 @@ class Pkcs12HelperUtils {
       P = Uint8List(0);
     }
     var I = Uint8List((S.length + P.length));
-    arrayCopy(S, 0, I, 0, S.length);
-    arrayCopy(P, 0, I, S.length, P.length);
+    _arrayCopy(S, 0, I, 0, S.length);
+    _arrayCopy(P, 0, I, S.length, P.length);
     var B = Uint8List(v);
     var c = (((n + u) - 1) ~/ u);
     var A = Uint8List(u);
@@ -49,25 +93,25 @@ class Pkcs12HelperUtils {
         B[j] = A[j % A.length];
       }
       for (var j = 0; j != (I.length ~/ v); j++) {
-        adjust(I, j * v, B);
+        _adjust(I, j * v, B);
       }
       if (i == c) {
-        arrayCopy(A, 0, dKey, (i - 1) * u, dKey.length - ((i - 1) * u));
+        _arrayCopy(A, 0, dKey, (i - 1) * u, dKey.length - ((i - 1) * u));
       } else {
-        arrayCopy(A, 0, dKey, (i - 1) * u, A.length);
+        _arrayCopy(A, 0, dKey, (i - 1) * u, A.length);
       }
     }
     return dKey;
   }
 
-  static void arrayCopy(Uint8List? sourceArr, int sourcePos, Uint8List? outArr,
+  void _arrayCopy(Uint8List? sourceArr, int sourcePos, Uint8List? outArr,
       int outPos, int len) {
     for (var i = 0; i < len; i++) {
       outArr![outPos + i] = sourceArr![sourcePos + i];
     }
   }
 
-  static void adjust(Uint8List a, int aOff, Uint8List b) {
+  void _adjust(Uint8List a, int aOff, Uint8List b) {
     var x = (b[b.length - 1] & 0xff) + (a[aOff + b.length - 1] & 0xff) + 1;
 
     a[aOff + b.length - 1] = x;
@@ -77,23 +121,6 @@ class Pkcs12HelperUtils {
       x += (b[i] & 0xff) + (a[aOff + i] & 0xff);
       a[aOff + i] = x;
       x >>>= 8;
-    }
-  }
-
-  static Uint8List formatPkcs12Password(String key) {
-    var password = key.codeUnits;
-    if (password.isNotEmpty) {
-      // +1 for extra 2 pad bytes.
-      var bytes = Uint8List((password.length + 1) * 2);
-
-      for (var i = 0; i != password.length; i++) {
-        bytes[i * 2] = (password[i] >>> 8);
-        bytes[i * 2 + 1] = password[i];
-      }
-
-      return bytes;
-    } else {
-      return Uint8List(0);
     }
   }
 }