+++ lib/src/model/asn1/pkcs/safe_bag.dart
@@ -1,8 +1,8 @@
 import 'dart:typed_data';
 
 import 'package:basic_utils/basic_utils.dart';
-import 'package:basic_utils/src/model/asn1/asn1/pkcs/cert_bag.dart';
-import 'package:basic_utils/src/model/asn1/asn1/pkcs/encrypted_private_key_info.dart';
+import 'package:basic_utils/src/model/asn1/pkcs/cert_bag.dart';
+import 'package:basic_utils/src/model/asn1/pkcs/encrypted_private_key_info.dart';
 
 ///
 ///```